backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.7.1
pymongo==4.10.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
fastapi-mail==1.4.2
email-validator==2.2.0
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
dnspython==2.7.0
